support multiple edge cycles

+++ firmware/include/physics/indexedEdge.hpp
@@ -9,8 +9,8 @@ class Obstacle;
 struct IndexedEdge
 {
     Obstacle* m_obstacle;
-    uint32_t m_index;
-    IndexedEdge(Obstacle* obstacle, uint32_t index);
+    uint16_t m_index;
+    IndexedEdge(Obstacle* obstacle, uint16_t index);
     bool operator==(const IndexedEdge& other) const;
     bool operator<(const IndexedEdge& other) const;
     Edge getEdge() const;
